moved saveAs function to utils.js
