+++ src/js/map/map_download.js
@@ -1,4 +1,4 @@
-import {Observable} from "../utils";
+import {Observable, saveAs} from "../utils";
 import html2canvas from 'html2canvas';
 
 export class MapDownload extends Observable {
@@ -19,31 +19,11 @@ export class MapDownload extends Observable {
             useCORS: true
         };
         let element = document.getElementById("main-map");
-        let self = this;
 
         html2canvas(element, options).then(function (canvas) {
-            self.saveAs(canvas.toDataURL(), 'map.png');
+            saveAs(canvas.toDataURL(), 'map.png');
         });
     }
 
-    saveAs = (uri, filename) => {
-        let link = document.createElement('a');
-        if (typeof link.download === 'string') {
 
-            link.href = uri;
-            link.download = filename;
-
-            //Firefox requires the link to be in the body
-            document.body.appendChild(link);
-
-            //simulate click
-            link.click();
-
-            //remove the link when done
-            document.body.removeChild(link);
-
-        } else {
-            window.open(uri);
-        }
-    }
 }
